backend/requirements.txt
aniso8601==9.0.1
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.1
flask-restx==1.3.0
Flask-Session==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
packaging==24.0
Pygments==2.17.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.0
restx==0.1.14
rich==13.7.1
rpds-py==0.18.0
shellingham==1.5.4
sniffio==1.3.1
typer==0.12.3
typing_extensions==4.11.0
Werkzeug==3.0.2

icecream~=2.1.3
pytest~=8.2.0
pydantic~=2.9.0
Flask-SocketIO~=5.3.6
SQLAlchemy~=2.0.36
Flask-Migrate~=4.0.7